+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add HTTP proxy for desktop ([#322](https://github.com/getsentry/sentry-unreal/pull/322))
 - Add scope support for Windows/Linux ([#328](https://github.com/getsentry/sentry-unreal/pull/328))
 - Add extra crash context for native integration ([#342](https://github.com/getsentry/sentry-unreal/pull/342))
+- Add missing plugin settings ([#335](https://github.com/getsentry/sentry-unreal/pull/335))
 
 ### Fixes
 
@@ -32,9 +33,9 @@
 - Bump Cocoa SDK (iOS) from v8.9.1 to v8.9.2 ([#331](https://github.com/getsentry/sentry-unreal/pull/331))
   - [changelog](https://github.com/getsentry/sentry-cocoa/blob/main/CHANGELOG.md#892)
   - [diff](https://github.com/getsentry/sentry-cocoa/compare/8.9.1...8.9.2)
-- Bump CLI from v2.19.4 to v2.20.1 ([#333](https://github.com/getsentry/sentry-unreal/pull/333), [#338](https://github.com/getsentry/sentry-unreal/pull/338))
-  - [changelog](https://github.com/getsentry/sentry-cli/blob/master/CHANGELOG.md#2201)
-  - [diff](https://github.com/getsentry/sentry-cli/compare/2.19.4...2.20.1)
+- Bump CLI from v2.19.4 to v2.20.3 ([#333](https://github.com/getsentry/sentry-unreal/pull/333), [#338](https://github.com/getsentry/sentry-unreal/pull/338), [#341](https://github.com/getsentry/sentry-unreal/pull/341))
+  - [changelog](https://github.com/getsentry/sentry-cli/blob/master/CHANGELOG.md#2203)
+  - [diff](https://github.com/getsentry/sentry-cli/compare/2.19.4...2.20.3)
 - Bump Cocoa SDK (iOS) from v8.9.2 to v8.9.3 ([#334](https://github.com/getsentry/sentry-unreal/pull/334))
   - [changelog](https://github.com/getsentry/sentry-cocoa/blob/main/CHANGELOG.md#893)
   - [diff](https://github.com/getsentry/sentry-cocoa/compare/8.9.2...8.9.3)